+++ src-tauri/src/dailies/point.rs
@@ -10,20 +10,20 @@ use crate::dailies::requirements::Requirements;
 pub struct Point {
     pub id: String,
     pub quest_id: String,
+    pub user_id: i64,
     pub date: NaiveDate,
     pub points: Option<f64>,
-    pub weight: f64,
     pub total: f64,
-    #[serde_as(as = "Option<i64>")]
+    pub complete: Option<f64>,
+    pub weight: f64,
+    pub points_weighted: Option<f64>,
     pub streak_target: Option<i64>,
+    pub streak: Option<i64>,
+    pub previous_streak: Option<i64>,
     pub requirements: Option<Requirements>,
-    #[serde_as(as = "Option<NaiveTime>")]
     pub time_start: Option<NaiveTime>,
-    #[serde_as(as = "Option<NaiveTime>")]
     pub time_end: Option<NaiveTime>,
-    #[serde_as(as = "Option<String>")]
     pub note: Option<String>,
-    #[serde_as(as = "NaiveDateTime")]
     pub updated: NaiveDateTime,
 }